requirements.txt
# Ad-Ops-Autopilot — Core dependencies
# Python 3.10+

# LLM
google-genai>=1.52.0,<2.0.0

# Data & visualization
pandas>=2.0.0,<3.0.0
matplotlib>=3.7.0,<4.0.0

# Token counting
tiktoken>=0.5.0,<0.8.0

# Config & env
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0,<7.0

# Testing
pytest>=7.4.0,<9.0.0

# Video generation (Fal.ai)
fal-client>=0.13.0,<1.0.0
httpx>=0.27.0,<1.0.0

# Linting
ruff>=0.1.0,<0.8.0
